SubnetTree: Update swig generated files
Re-generate SubnetTree.py and SubnetTree_wrap.cc using the
command in the Makefile with swig 3.0.12 (rather than newer 4.0)
to keep changes minimal.

The added tests fail as follows with the current version installable
from pip:

    $ btest -d -v pysubnettree/errors.test
    pysubnettree.errors ...
      > python3 %INPUT
    ... pysubnettree.errors failed
      % 'python3
    <...>/.tmp/pysubnettree.errors/errors.test'
    failed unexpectedly (exit code 139)
      % cat .stderr
      Segmentation fault (core dumped)

This requires a 0.36 release and upload to pypi to fix pip users.

I'm optimistically marking the following issues as closed (once we
do the release they should actually be fixed).

Closes #33
Closes #32
Closes #15